Renamed badly named cldr_names to language_name_db for better readability
