+++ suggestedimages/locales/locales/es.py
@@ -1,4 +1,4 @@
-from ..cldr_names import LanguageNames
+from ..language_name_db import LanguageNames
 
 texts = {
     '[no label]': '[sin etiqueta]',
